// client-next/pages/main/profile/types/profile-types.ts
export type ChangeInfo = ({id, text}: TextProps) => Promise<void>
export type PostType = {
    text: string
    date: Date
    id: number
}

export type TextProps = {
    text: string
    id: number
}

export interface ProfileInterface {
    name: string
    location: string
    banner: any
    avatar: any
}